requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
alembic==1.4.2
    # via -r requirements.in
anytree==2.8.0
    # via -r requirements.in
apispec==0.39.0
    # via -r requirements.in
atomicwrites==1.4.0
    # via -r requirements.in
blinker==1.5
    # via -r requirements.in
boltons==23.0.0
    # via -r requirements.in
cairocffi==1.4.0
    # via
    #   cairosvg
    #   weasyprint
cairosvg==2.5.2
    # via weasyprint
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via
    #   cairocffi
    #   weasyprint
charset-normalizer==2.0.12
    # via requests
click==6.7
    # via flask
click-spinner==0.1.8
    # via -r requirements.in
colorama==0.3.9
    # via -r requirements.in
colour==0.1.5
    # via -r requirements.in
cssselect2==0.7.0
    # via
    #   cairosvg
    #   weasyprint
defusedxml==0.7.1
    # via
    #   cairosvg
    #   odfpy
et-xmlfile==1.1.0
    # via
    #   -r requirements.in
    #   openpyxl
flask==1.0.2
    # via
    #   flask-cors
    #   flask-login
    #   flask-sqlalchemy
    #   flask-weasyprint
    #   flask-wtf
flask-cors==3.0.10
    # via -r requirements.in
flask-login==0.5.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
flask-weasyprint==0.4
    # via -r requirements.in
flask-wtf==1.0.0
    # via -r requirements.in
hashids==1.2.0
    # via -r requirements.in
html5lib==1.1
    # via weasyprint
idna==3.4
    # via requests
inflection==0.5.1
    # via -r requirements.in
itsdangerous==2.0.1
    # via
    #   flask
    #   flask-wtf
jinja2==3.0.3
    # via flask
mako==1.2.3
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
marshmallow==3.0.0b11
    # via
    #   -r requirements.in
    #   marshmallow-enum
    #   webargs
marshmallow-enum==1.4.1
    # via -r requirements.in
more-itertools==8.12.0
    # via -r requirements.in
numpy==1.22.0
    # via
    #   -r requirements.in
    #   pandas
odfpy==1.4.1
    # via -r requirements.in
openpyxl==3.0.10
    # via -r requirements.in
pandas==1.3.5
    # via -r requirements.in
passlib==1.7.1
    # via -r requirements.in
phonenumbers==8.9.11
    # via -r requirements.in
pillow==9.2.0
    # via cairosvg
pint==0.9
    # via -r requirements.in
psycopg2-binary==2.8.3
    # via -r requirements.in
py-dmidecode==0.1.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via -r requirements.in
pyphen==0.13.0
    # via weasyprint
python-dateutil==2.7.3
    # via
    #   alembic
    #   pandas
python-decouple==3.3
    # via -r requirements.in
python-dotenv==0.14.0
    # via -r requirements.in
python-editor==1.0.4
    # via alembic
python-stdnum==1.9
    # via -r requirements.in
pytz==2022.2.1
    # via pandas
pyyaml==5.4
    # via
    #   -r requirements.in
    #   apispec
requests==2.27.1
    # via
    #   -r requirements.in
    #   requests-mock
    #   requests-toolbelt
requests-mock==1.5.2
    # via -r requirements.in
requests-toolbelt==0.9.1
    # via -r requirements.in
six==1.16.0
    # via
    #   anytree
    #   flask-cors
    #   html5lib
    #   python-dateutil
    #   requests-mock
    #   sqlalchemy-utils
sortedcontainers==2.1.0
    # via -r requirements.in
sqlalchemy==1.3.24
    # via
    #   alembic
    #   flask-sqlalchemy
    #   sqlalchemy-citext
    #   sqlalchemy-utils
sqlalchemy-citext==1.3.post0
    # via -r requirements.in
sqlalchemy-utils==0.33.11
    # via -r requirements.in
tinycss2==1.1.1
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
tqdm==4.32.2
    # via -r requirements.in
urllib3==1.26.12
    # via requests
weasyprint==44
    # via flask-weasyprint
webargs==5.5.3
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
werkzeug==2.0.3
    # via
    #   -r requirements.in
    #   flask
wtforms==3.0.1
    # via flask-wtf
xlrd==2.0.1
    # via -r requirements.in
